Cleared SITE_DOMAIN.
Keeping SITE_DOMAIN empty results in a dynamic script, plus it works
out of the box.
That's possible because every link in CSS, HTML, ... without "http://"
prefix is relative to the main document, index.php in this case; the
browser is using the same domain and port information it used to access
index.php.