Update: index.php and field.js refactoring
No longer generating JavaScript from the saved lineup.txt files
in index.php. Instead, we are now creating the player divs with
data-label to hold the player names. I figured that would be a
bit safer.

field.js modified as a consequence.

Also, added a .gitignore rule for the lineups directory.